+++ Assig1/ViewModels/Expiation_ExpiationDetail.cs
@@ -7,6 +7,6 @@ namespace Assig1.ViewModels
         public Expiation Expiation { get; set; }
         public int DriverCount { get; set; }
         public int LsaCount { get; set; }
-        public int SpeedPercentile { get; set; }
+        public double SpeedPercentile { get; set; }
     }
 }
